requirements.txt

django-environ==0.11.2
psycopg2==2.9.9
python-jose==3.3.0

fastapi==0.103.1
pydantic==2.3.0
pydantic-settings==2.0.3
uvicorn==0.23.2